refactor: 노선 역 목록 조회 로직을 도메인으로 분리
